Extensive refactoring of the configuration file, to move all environment-specific stuff to a top-level key: "env".
I have re-coded a bunch of stuff to comply with the new config tree structure.  But I still have to remove the hard-coded environment constants.

Also, the "-dev" arg for setting the build environment to "dev" has been replaced with a more general config-setting scheme: "-DbuildEnv=dev".  Any top-level key is supported in place of "buildEnv".